rename test_env.py to test_environment.py
