+++ tests/wrappers/test_action_by_acceleration.py
@@ -9,7 +9,7 @@ from pynktrombonegym.environment import PynkTrombone
 from pynktrombonegym.spaces import ActionSpaceNames as ASN
 from pynktrombonegym.wrappers.action_by_acceleration import ActionByAcceleration
 
-from ..test_env import assert_dict_space_key, assert_space, target_sound_files
+from ..test_environment import assert_dict_space_key, assert_space, target_sound_files
 from ..test_generation import generate_sound
 
 initial_pos = OrderedDict(
